[C] Rename ArticleMetadataBlock to EntityMetadataBlock
