+++ packages/frontend/components/factories/EntityMetadataFactory/patterns/Paper.tsx
@@ -1,7 +1,7 @@
 import { graphql } from "react-relay";
 import { useTranslation } from "react-i18next";
 import { useMaybeFragment } from "@wdp/lib/api/hooks";
-import ArticleMetadataBlock from "components/composed/article/ArticleMetadataBlock/ArticleMetadataBlock";
+import EntityMetadataBlock from "components/composed/entity/EntityMetadataBlock/EntityMetadataBlock";
 import { MetadataProperty } from "components/layout";
 import MetadataFactory from "components/factories/MetadataFactory";
 import ContributorName from "components/composed/contributor/ContributorName";
@@ -21,7 +21,7 @@ export default function PaperMetadata({ data }: Props) {
   );
 
   return paper ? (
-    <ArticleMetadataBlock>
+    <EntityMetadataBlock>
       <MetadataProperty
         label={t("metadata.author", { count: authors?.length ?? 1 })}
       >
@@ -47,7 +47,7 @@ export default function PaperMetadata({ data }: Props) {
         label={t("metadata.pdf_version")}
         data={paper.pdfVersion}
       />
-    </ArticleMetadataBlock>
+    </EntityMetadataBlock>
   ) : null;
 }
