+++ packages/frontend/components/factories/EntityMetadataFactory/patterns/parts/ArticleIssueMetadata.tsx
